msvc64: fixes to avoid build warnings (which are treated as errors) from long/size_t mismatches.

+++ gsoc13/src/basicio.cpp
@@ -2079,7 +2079,7 @@ namespace Exiv2 {
         byte* data = (byte*) std::malloc(dataSize);
         src.seek(left, BasicIo::beg);
         src.read(data, dataSize);
-        p_->httpPost(data, dataSize, left, p_->size_ - right);
+        p_->httpPost(data, dataSize, left, (long) p_->size_ - right);
         if (data) std::free(data);
         return src.size();
     }
